requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
python-dotenv
requests
openpyxl  